Rename `requests/` directory to `graphql/`
I'm going to create actual request functions in the requests directory, which use the graphql queries appropriately. Best to have all this done in one place!
