+++ src/components/templates/Blog/get-article-index.ts
@@ -1,4 +1,4 @@
-import {indexArticles} from '@helpers/requests/articles'
+import {indexArticles} from '@helpers/graphql/articles'
 import {request} from 'graphql-request'
 
 const getArticleIndex = async ( ) => {
